fixed importer fixed several compiler warnings while rebuilding the project (this takes soo long :/ ) fixed math lib
git-svn-id: svn+ssh://131.220.23.26/daten/v3d/trunk@558 165746f0-bcf8-0310-b0ad-99815d7186ab

+++ velox/Source/Resource/Types/VMaterialResourceType.cpp
@@ -99,7 +99,7 @@ VRenderStateList* VMaterialResourceType::CreatePass(const VRenderPass& in_Pass) 
 
 vbool VMaterialResourceType::IsCGFXFile(const std::string& in_strFileName) const
 {
-	const int len = in_strFileName.length();
+	const int len = static_cast<int>(in_strFileName.length());
 
 	if( len > 2 && in_strFileName[len-2] == 'f' && in_strFileName[len-1] == 'x' )
 		return true;
